Move discovery functions to elastic.py to reduce duplication
