+++ src/prime_rl/orchestrator/env_worker.py
@@ -13,13 +13,12 @@ from itertools import cycle
 from multiprocessing import Process, Queue
 from pathlib import Path
 
-import httpx
 import verifiers as vf
 from openai import AsyncOpenAI
 
 from prime_rl.utils.client import setup_clients
 from prime_rl.utils.config import ClientConfig
-from prime_rl.utils.elastic import discover_server_ips
+from prime_rl.utils.elastic import discover_ready_servers
 from prime_rl.utils.logger import get_logger, intercept_verifiers_logging, reset_logger, setup_logger
 
 
@@ -46,47 +45,6 @@ class RolloutResponse:
     request_id: str
     results: list[dict]
     lag_metrics: dict | None = None
-
-
-async def _check_server(url: str, model_name: str, timeout: float = 5.0) -> tuple[bool, bool]:
-    """Check server status. Returns (has_model, is_healthy)."""
-    try:
-        async with httpx.AsyncClient(timeout=timeout) as client:
-            response = await client.get(f"{url}/v1/models")
-            response.raise_for_status()
-            data = response.json()
-            models = [m.get("id") for m in data.get("data", [])]
-            return model_name in models, len(models) > 0
-    except Exception:
-        return False, False
-
-
-async def discover_ready_servers(hostname: str, port: int, model_name: str) -> list[str]:
-    """Discover servers via DNS with majority vote logic.
-
-    - If NO servers have the model: return all healthy servers (base model mode)
-    - If ANY server has the model: return only those with it (adapter mode)
-    """
-    loop = asyncio.get_event_loop()
-    ips = await loop.run_in_executor(None, discover_server_ips, hostname)
-    if not ips:
-        return []
-
-    checks = [_check_server(f"http://{ip}:{port}", model_name) for ip in ips]
-    results = await asyncio.gather(*checks, return_exceptions=True)
-
-    with_model, healthy = [], []
-    for ip, result in zip(ips, results):
-        if isinstance(result, Exception):
-            continue
-        has_model, is_healthy = result
-        url = f"http://{ip}:{port}/v1"
-        if has_model:
-            with_model.append(url)
-        if is_healthy:
-            healthy.append(url)
-
-    return with_model if with_model else healthy
 
 
 def extract_result(state: vf.State) -> dict:
